refactor(openings): split OpeningsPage into hooks + tab subcomponents
Behavior-preserving extraction. The original ~1843-line OpeningsPage
mixed deep-link highlight state, sidebar/drawer state, tab reset state,
deep-link navigation handlers, and four tabs of JSX into a single
component.

Extracted under frontend/src/pages/openings/:
  - useDeepLinkHighlight  — owns highlightedMove + pulse-active state,
    cross-tab clear, and filter-change clear effects.
  - useSidebarState       — owns desktop sidebar panel + mobile drawer
    open state + persisted onboarding hint flags.
  - useTabReset           — owns gamesOffset + prevTab tracking; resets
    pagination on tab switch.
  - useOpeningsHandlers   — bundles deep-link / open-from-{X} navigation
    callbacks (handleOpenChartBookmarkGames, handleOpenGames,
    handleOpenMoves, handleOpenFinding, handleOpenFindingGames,
    handleLoadBookmark, handleReorder).
  - ExplorerTab.tsx       — Move Explorer subtab JSX (WDL/Score/Eval
    bullet rows + MoveExplorer).
  - GamesTab.tsx          — Games subtab JSX (loading / no-games /
    no-match / error / list states).
  - StatsTab.tsx          — Stats subtab JSX (bookmark + most-played
    sections, ScoreChart over time).
  - InsightsTab.tsx       — Insights subtab JSX (OpeningInsightsBlock).

The shell OpeningsPage retains the chessboard column (shared across
tabs), the desktop sidebar/strip + mobile drawer chrome, the routing/
tab-state, and the data-fetching hook calls. Per-tab queries stay in
the shell to preserve the original timing semantics (queries fire on
parent mount, not on tab activation).

All data-testid values (57 unique) and aria-label / ariaLabel props are
preserved verbatim. Mobile drawer + desktop sidebar render paths are
both kept intact and routed through the same hook state.

No behavior changes. No new public exports. No query refactors. Lint
(eslint + knip), tests (311/311), and production build all green.
